chore(cli): refactor download into modules
